WIP: api constants transferred to correct file
