+++ src/constants/api.js
@@ -1,0 +1,80 @@
+export const RESOURCES = {
+  RATE: 'rates',
+  CDR: 'cdrs',
+  AUTH_ORIGIN_TRANSPORT_PROTOCOL: 'authOrigTransportProtocol',
+  ACCOUNT: 'accounts',
+  AUTH: 'auth',
+};
+
+export const RELATIONSHIPS = {
+  [RESOURCES.AUTH]: {
+    login: '',
+    password: '',
+    jwt: '',
+  },
+  [RESOURCES.RATE]: {
+    prefix: '',
+    'initial-rate': '',
+    'initial-interval': '',
+    'next-rate': '',
+    'next-interval': '',
+    'connect-fee': '',
+    'reject-calls': '',
+    'valid-from': '',
+    'valid-till': '',
+    'network-prefix-id': '',
+  },
+  [RESOURCES.CDR]: {
+    'time-start': '',
+    'time-connect': '',
+    'time-end': '',
+    duration: '',
+    success: '',
+    'destination-initial-interval': '',
+    'destination-initial-rate': '',
+    'destination-next-interval': '',
+    'destination-next-rate': '',
+    'destination-fee': '',
+    'customer-price': '',
+    'src-name-in': '',
+    'src-prefix-in': '',
+    'from-domain': '',
+    'dst-prefix-in': '',
+    'to-domain': '',
+    'ruri-domain': '',
+    'diversion-in': '',
+    'local-tag': '',
+    'lega-disconnect-code': '',
+    'lega-disconnect-reason': '',
+    'lega-rx-payloads': '',
+    'lega-tx-payloads': '',
+    'auth-orig-transport-protocol-id': '',
+    'auth-orig-ip': '',
+    'auth-orig-port': '',
+    'lega-rx-bytes': '',
+    'lega-tx-bytes': '',
+    'lega-rx-decode-errs': '',
+    'lega-rx-no-buf-errs': '',
+    'lega-rx-parse-errs': '',
+    'src-prefix-routing': '',
+    'dst-prefix-routing': '',
+    'destination-prefix': '',
+    'auth-orig-transport-protocol': {
+      jsonApi: 'hasMany',
+      type: 'comments',
+    },
+  },
+  [RESOURCES.AUTH_ORIGIN_TRANSPORT_PROTOCOL]: {
+    'auth-orig-transport-protocol': 'auth-orig-transport-protocol',
+  },
+  [RESOURCES.ACCOUNT]: {
+    name: '',
+    balance: '',
+    'min-balance': '',
+    'max-balance': '',
+    'destination-rate-limit': '',
+    'origination-capacity': '',
+    'termination-capacity': '',
+    'total-capacity': '',
+  },
+};
