+++ src/store/modules/cdrs.js
@@ -1,7 +1,6 @@
 // eslint-disable-next-line
 import { jsonApi } from '../../api';
-import { RESOURCES } from '../../static/constants/api';
-import { CDRS } from '../../constants';
+import { RESOURCES, CDRS } from '../../constants';
 
 const state = {
   cdrs: {},
